Revert "Commit 2 error fix attempt 2"
This reverts commit 6b9b7f8ec8b7f2eb95bcb1287c637af63e517494.
